+++ app/reward-economy/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import RewardEconomyDashboard from "@/components/reward-economy/RewardEconomyDashboard";
+
+export default function RewardEconomyPage() {
+  return <RewardEconomyDashboard />;
+}